rename and format some files
